% Generated by roxygen2 (4.0.1): do not edit by hand
\name{combined.error}
\alias{combined.error}
\title{combined.error}
\usage{
combined.error(df1, df2, categorical.inds)
}
\description{
I measure categorical vairable imputation error as absolute error
and continuous variable imputation error as mean squared error
}

