% Generated by roxygen2 (4.0.1): do not edit by hand
\name{continuous.error}
\alias{continuous.error}
\title{I measure the mean squared error of imputing the ontinuous variables}
\usage{
continuous.error(df1, df2)
}
\description{
I measure the mean squared error of imputing the ontinuous variables
}

