\name{fastimputer-package}
\alias{fastimputer-package}
\alias{fastimputer}
\docType{package}
\title{
Really Fast Imputation
}
\description{
A package with options to fill in missing values using simple, fast methods.
Also, offers pre- and post- imputation analysis of time, accuracy and missingness.
}
\details{
\tabular{ll}{
Package: \tab fastimputer\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-07-13\cr
License: \tab MIT\cr
}
todo
}
\author{
john collins
Maintainer: john collins <johnssocks@gmail.com>
}
\references{
none}
impute, imputation
documentation directory
\keyword{ package }
\seealso{

}
\examples{
}
