% Generated by roxygen2 (4.0.1): do not edit by hand
\name{preimputation.analyzer}
\alias{preimputation.analyzer}
\title{preimputation.analyzer}
\usage{
preimputation.analyzer(df)
}
\arguments{
\item{df}{Data frame for all data including missing and non-missing values}
}
\description{
I give a user the option to predict how long using
this module will take and how accurate the results
will be by doing imputation on a small set and
extrapolating. Further, I give advice about what
kinds of missingness are observed within the data.
}
\examples{
df1 <- data.frame(A=c(1,2,3,1,2,3,1,2,3,1,2,3,1,2,3), B=as.factor(c(1,2,1,3,NA,2,1,2,3,NA,1,3,1,2,NA)), C=c(1.1, 3.5, NA, 3, NA, NA, 4.1, 2.2, 2.3, 1.4, 1, 2, 5, 3, 4))
preimputation.analysis <- preimputation.analyzer(df1)
df2 <- read.table(
"http://www.unt.edu/rss/class/Jon/Benchmarks/rrp.ex.data.txt",
header = TRUE, sep = ",", na.strings = "NA",
dec = ".", strip.white = TRUE)
preimputation.analysis <- preimputation.analyzer(df2)
}
\keyword{imputation,}
\keyword{impute,}
\keyword{imputer,}
\keyword{missing}

